metric improvements and reverted cmakelists.txt

+++ include/simulator/adc/adc.h
@@ -25,8 +25,8 @@ class ADC : public Component {
     int calc_transistor_count() override;
     double calc_static() override;
     double calc_dynamic() override;
-    int calc_width() override;
-    int calc_height() override;
+    double calc_width() override;
+    double calc_height() override;
 
     void update_static(double time) override;
     int get_cycle_count() override;
