+++ include/simulator/metrics/packer.h
@@ -11,11 +11,11 @@
 class Packer {
     std::shared_ptr<PackNode> root_ = std::make_shared<PackNode>();
 
-    std::shared_ptr<PackNode> find_node(std::shared_ptr<PackNode> root, int width, int height);
-    std::shared_ptr<PackNode> grow_node(int width, int height);
-    std::shared_ptr<PackNode> grow_right(int width, int height);
-    std::shared_ptr<PackNode> grow_down(int width, int height);
-    std::shared_ptr<PackNode> split_node(std::shared_ptr<PackNode> node, int width, int height);
+    std::shared_ptr<PackNode> find_node(std::shared_ptr<PackNode> root, double width, double height);
+    std::shared_ptr<PackNode> grow_node(double width, double height);
+    std::shared_ptr<PackNode> grow_right(double width, double height);
+    std::shared_ptr<PackNode> grow_down(double width, double height);
+    std::shared_ptr<PackNode> split_node(std::shared_ptr<PackNode> node, double width, double height);
 
    public:
     std::shared_ptr<PackNode> pack(std::vector<std::shared_ptr<Component> >& components);
