+++ include/simulator/registers/register.h
@@ -47,8 +47,8 @@ class Register : public Component {
     cv::Mat get_transistor_count_array() override;
     cv::Mat get_static_energy_array() override;
     cv::Mat get_dynamic_energy_array() override;
-    int get_width() override;
-    int get_height() override;
+    double get_width() override;
+    double get_height() override;
     int get_transistor_count() override;
 #endif
