+++ src/simulator/registers/register.cpp
@@ -141,13 +141,13 @@ void Register::inc_write() {
     }
 }
 
-int Register::get_width() {
+double Register::get_width() {
     if (memory_) {
         return this->memory_->get_width();
     }
 }
 
-int Register::get_height() {
+double Register::get_height() {
     if (memory_) {
         return this->memory_->get_height();
     }
